requirements.txt
sqlalchemy
sqlalchemy-utils
psycopg2-binary
passlib==1.7.4
bcrypt==4.0.1
flask
flask-cors
python-jose[cryptography]
python-dotenv
cryptography
flask-limiter
alembic
pytest
pytest-cov
factory-boy
pytest-flask
colorama
pillow
qrcode